fix : LogoutController.java 삭제 및 AuthController에 logoutAPI추가
