+++ src/main/java/com/project/likelion14thbe/domain/auth/controller/AuthController.java
@@ -4,6 +4,7 @@ import com.project.likelion14thbe.domain.auth.service.AuthService;
 import com.project.likelion14thbe.global.apiPayload.exception.CustomResponse;
 import com.project.likelion14thbe.global.security.jwt.dto.JwtDTO;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -29,5 +30,18 @@ public class AuthController {
         log.info("[ Auth Controller ] 토큰을 재발급합니다. ");
 
         return CustomResponse.onSuccess(authService.reissueToken(jwtDto));
+    }
+
+    // 로그아웃 API (Swagger 문서화 전용 - 실제 처리는 Security LogoutFilter)
+    @Operation(
+            summary = "로그아웃",
+            description = "Authorization 헤더에 Bearer {accessToken} 을 담아 호출합니다. " +
+                    "DB의 Refresh Token 을 삭제하고, Access Token 을 블랙리스트에 등록하며 SecurityContext 를 클리어합니다.",
+            security = @SecurityRequirement(name = "JWT TOKEN")
+    )
+    @PostMapping("/logout")
+    public CustomResponse<String> logout() {
+        // LogoutFilter 가 이 메서드 호출 전에 가로채므로 실제로 실행되지 않음
+        return CustomResponse.onSuccess("로그아웃이 완료되었습니다.");
     }
 }
